move ocw tests
